# services/audit-compliance-service/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Audit Compliance Service
  version: 0.1.0
  description: |
    Hand-authored 1:1 slice of the audit-compliance-service HTTP surface used
    as the canonical input for `tools/of-sdk-gen`. The endpoints listed here
    mirror the chi router wiring in `internal/server/server.go` (audit ledger
    subset). This is intentionally a small POC slice — not the full surface.
servers:
  - url: http://localhost:7141
    description: Local default
...
paths:
  /api/v1/audit/events:
    post:
      summary: Append an audit event
      operationId: appendAuditEvent
      description: Anonymous append; gateway / in-cluster callers only.
      tags: [audit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendAuditEventRequest'
      responses:
        '201':
          description: Event appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '400':
          description: Invalid body
    get:
      summary: List audit events
      operationId: listAuditEvents
      tags: [audit]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: source_service
          schema: { type: string }
        - in: query
          name: subject_id
          schema: { type: string }
        - in: query
          name: classification
          schema: { type: string }
        - in: query
          name: resource_id
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: trace_id
          schema: { type: string }
        - in: query
          name: event_id
          schema: { type: string }
      responses:
        '200':
          description: Page of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
  /api/v1/audit/events/{id}:
    get:
      summary: Get an audit event by id
      operationId: getAuditEvent
      tags: [audit]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AppendAuditEventRequest:
      type: object
      required: [source_service, channel, actor, action, resource_type, resource_id, status, severity, classification]
      properties:
        source_service:   { type: string }
        product:          { type: string }
        product_version:  { type: string }
        producer_type:    { type: string }
        channel:          { type: string }
        event_id:         { type: string, format: uuid }
        log_entry_id:     { type: string, format: uuid }
        sequence_id:      { type: string, format: uuid, nullable: true }
        actor:            { type: string }
        actor_id:         { type: string }
        actor_type:       { type: string }
        session_id:       { type: string, nullable: true }
        service_account_id: { type: string, nullable: true }
        token_id:         { type: string, nullable: true }
        action:           { type: string }
        categories:
          type: array
          items: { type: string }
        resource_type:    { type: string }
        resource_id:      { type: string }
        entities:
          type: array
          items:
            type: object
            additionalProperties: true
        origins:
          type: array
          items: { type: string }
        origin:           { type: string, nullable: true }
        source_origin:    { type: string, nullable: true }
        trace_id:         { type: string, nullable: true }
        status:
          type: string
          enum: [success, failure, denied]
        outcome:
          type: string
          enum: [success, error, unauthorized]
        severity:
          type: string
          enum: [low, medium, high, critical]
        classification:   { type: string }
        subject_id:       { type: string, nullable: true }
        ip_address:       { type: string, nullable: true }
        location:         { type: string, nullable: true }
        metadata:         { type: object, additionalProperties: true }
        error_metadata:   { type: object, additionalProperties: true }
        request_fields:   { type: object, additionalProperties: true }
        result_fields:    { type: object, additionalProperties: true }
        labels:
          type: array
          items: { type: string }
        parent_event_id:  { type: string, format: uuid, nullable: true }
        initiator_type:   { type: string }
        audit_access_tier: { type: string }
        retention_days:
          type: integer
          format: int32
          minimum: 0
    AuditEvent:
      type: object
      required:
        - id
        - sequence
        - previous_hash
        - entry_hash
        - source_service
        - channel
        - actor
        - action
        - resource_type
        - resource_id
        - status
        - severity
        - classification
        - retention_until
        - occurred_at
        - ingested_at
      properties:
        id:             { type: string, format: uuid }
        event_id:       { type: string, format: uuid }
        log_entry_id:   { type: string, format: uuid }
        sequence_id:    { type: string, format: uuid, nullable: true }
        sequence:       { type: integer, format: int64 }
        previous_hash:  { type: string }
        entry_hash:     { type: string }
        source_service: { type: string }
        product:        { type: string }
        product_version: { type: string }
        producer_type:  { type: string }
        channel:        { type: string }
        actor:          { type: string }
        actor_id:       { type: string }
        actor_type:     { type: string }
        session_id:     { type: string, nullable: true }
        service_account_id: { type: string, nullable: true }
        token_id:       { type: string, nullable: true }
        action:         { type: string }
        categories:
          type: array
          items: { type: string }
        resource_type:  { type: string }
        resource_id:    { type: string }
        entities:
          type: array
          items:
            type: object
            additionalProperties: true
        origins:
          type: array
          items: { type: string }
        origin:         { type: string, nullable: true }
        source_origin:  { type: string, nullable: true }
        trace_id:       { type: string, nullable: true }
        status:         { type: string }
        outcome:        { type: string }
        severity:       { type: string }
        classification: { type: string }
        subject_id:     { type: string, nullable: true }
        ip_address:     { type: string, nullable: true }
        location:       { type: string, nullable: true }
        metadata:       { type: object, additionalProperties: true }
        error_metadata: { type: object, additionalProperties: true }
        request_fields: { type: object, additionalProperties: true }
        result_fields:  { type: object, additionalProperties: true }
        labels:
          type: array
          items: { type: string }
        parent_event_id: { type: string, format: uuid, nullable: true }
        initiator_type: { type: string }
        audit_access_tier: { type: string }
        retention_until: { type: string, format: date-time }
        occurred_at:     { type: string, format: date-time }
        ingested_at:     { type: string, format: date-time }
    AnomalyAlert:
      type: object
      required: [id, kind, severity, detected_at]
      properties:
        id:           { type: string, format: uuid }
        kind:         { type: string }
        severity:     { type: string }
        detected_at:  { type: string, format: date-time }
        description:  { type: string }
    EventListResponse:
      type: object
      required: [items, anomalies]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuditEvent' }
        anomalies:
          type: array
          items: { $ref: '#/components/schemas/AnomalyAlert' }
...